Time: 0 ms (100%), Space: 8.3 MB (8.32%) - LeetHub

+++ 1344-angle-between-hands-of-a-clock/1344-angle-between-hands-of-a-clock.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
     double angleClock(int hour, int minutes) {
-       float  mindeg=6.00*minutes;
+       double  mindeg=6.00*minutes;
      
-       float hrdg=30.00*hour;
+       double hrdg=30.00*hour;
         if (hour==12) hrdg=0;
 
-        float extramovehr=0.5*minutes;
+        double extramovehr=0.5*minutes;
 
-       float angle= abs(hrdg+extramovehr-mindeg);
+       double angle= abs(hrdg+extramovehr-mindeg);
        return min(angle, 360-angle);
     }
 };
